use the import syntax in server/termdb* file
