--- server/src/termdb.sql.cohort.js
@@ -1,12 +1,0 @@
-export function getCohortsData(ds) {
-	const features = ds.cohort.db.connection.prepare('select * from features').all()
-	const cohorts = ds.cohort.db.connection
-		.prepare(
-			`select * from cohorts where cohort in (select distinct(cohort) from cohort_features)
-		 order by sample_count desc`
-		)
-		.all()
-	const cfeatures = ds.cohort.db.connection.prepare('select * from cohort_features').all()
-
-	return { cohorts, features, cfeatures }
-}
